@@ -20,29 +20,30 @@
   "dependencies": {
     "@angular-redux/router": "^7.0.0",
     "@angular-redux/store": "^7.1.1",
-    "@angular/animations": "5.2.9",
-    "@angular/cdk": "^5.2.5",
-    "@angular/common": "5.2.9",
-    "@angular/compiler": "5.2.9",
-    "@angular/core": "5.2.9",
-    "@angular/flex-layout": "^5.0.0-beta.14",
-    "@angular/forms": "5.2.9",
-    "@angular/http": "5.2.9",
-    "@angular/material": "^5.2.5",
-    "@angular/platform-browser": "5.2.9",
-    "@angular/platform-browser-dynamic": "5.2.9",
-    "@angular/router": "5.2.9",
+    "@angular/animations": "6.0.0",
+    "@angular/cdk": "^6.0.0",
+    "@angular/common": "6.0.0",
+    "@angular/compiler": "6.0.0",
+    "@angular/core": "6.0.0",
+    "@angular/flex-layout": "^6.0.0-beta.15",
+    "@angular/forms": "6.0.0",
+    "@angular/http": "6.0.0",
+    "@angular/material": "^6.0.0",
+    "@angular/platform-browser": "6.0.0",
+    "@angular/platform-browser-dynamic": "6.0.0",
+    "@angular/router": "6.0.0",
     "core-js": "^2.4.1",
     "hammerjs": "^2.0.8",
     "redux": "^3.7.2",
     "redux-logger": "^3.0.6",
-    "rxjs": "^5.5.8",
+    "rxjs": "6.1.0",
     "zone.js": "^0.8.25"
   },
   "devDependencies": {
-    "@angular/cli": "1.7.4",
-    "@angular/compiler-cli": "5.2.9",
-    "@angular/language-service": "5.2.9",
+    "@angular-devkit/build-angular": "~0.6.0",
+    "@angular/cli": "^6.0.0",
+    "@angular/compiler-cli": "6.0.0",
+    "@angular/language-service": "6.0.0",
     "@types/jasmine": "^2.8.5",
     "@types/jasminewd2": "~2.0.3",
     "@types/node": "~6.0.60",
@@ -62,7 +63,7 @@
     "protractor": "~5.1.2",
     "ts-node": "~4.1.0",
     "tslint": "~5.9.1",
-    "typescript": "2.6.2"
+    "typescript": "2.7.2"
   },
   "lint-staged": {
     "*.{scss,js,css}": [
